% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_bank.R
\docType{data}
\name{bank}
\alias{bank}
\title{Bank Queue Data}
\format{
A data frame with 100 rows and 2 columns for arrival time and service time.
}
\source{
Mention where you got the data, if applicable.
}
\usage{
data(bank)
}
\description{
A dataset containing example bank queue data.
}
