% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Multiserver Queue System Simulation}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{A vector of customer arrival times.}

\item{ServiceTimes}{A vector of service times for each customer.}

\item{NumServers}{The number of servers available in the queue system. Default is 1.}
}
\value{
A tibble with columns: Arrivals, ServiceBegins, ChosenServer, and ServiceEnds.
}
\description{
Simulate a multi-server queue system to determine each customer's transition times.
}
\examples{
# Example usage:
Arrivals <- c(1, 3, 5)
ServiceTimes <- c(2, 2, 2)
Multiserver(Arrivals, ServiceTimes, 2)

}
